Move authorization to resolve_filter

+++ lib/graphoid/mutations/delete.rb
@@ -40,9 +40,9 @@ module Graphoid
         type.class_eval do
           define_method :"#{plural}" do |where: {}|
             begin
-              model.authorize_many!(context[:current_user]) if model.respond_to?(:authorize_many!)
               objects = Graphoid::Queries::Processor.execute(model, where.to_h)
               objects = model.resolve_filter(self, objects) if model.respond_to?(:resolve_filter)
+              raise objects if objects.is_a? GraphQL::ExecutionError
               objects.destroy_all
               objects.all.to_a
             rescue Exception => ex
